Print __PRETTY_FUNCTION__ of test function not helper lambda
On test failure, the COMPARE and ULP_COMPARE functions use
__PRETTY_FUNCTION__ to print the function where the failure happened.
As an implementation detail the macros internally use a directly invoked
lambda. If __PRETTY_FUNCTION__ is inside that lambda then the compiler
may emit the name of the lambda instead of the calling function. To get
to see the latter, pass __PRETTY_FUNCTION__ as function argument to the
lambda.

ChangeLog:

	* testsuite/tests/bits/verify.h (COMPARE, ULP_COMPARE): Move
	__PRETTY_FUNCTION__ from inside the lambda to the lambda
	arguments.